[feat] #37 Create sub package in ui and Divide screen according to step
